qw: add broken-ass port

+++ qw/d_fill.c
@@ -5,7 +5,6 @@
 #include <stdio.h>
 #include "quakedef.h"
 
-
 /*
 ================
 D_FillRect
@@ -42,7 +41,7 @@ void D_FillRect (vrect_t *rect, int color)
 
 	dest = ((byte *)vid.buffer + ry*vid.rowbytes + rx);
 
-	if (((rwidth & 0x03) == 0) && (((long)dest & 0x03) == 0))
+	if (((rwidth & 0x03) == 0) && (((uintptr)dest & 0x03) == 0))
 	{
 	// faster aligned dword clear
 		ldest = (unsigned *)dest;
